Add support for ssh-based Github repositories.
I noticed that the previous script only supported https-based github
repository. This commit adds the same functionality for ssh repos. It is
transparent for the user. This means they don't need to know if they are
on https or ssh, the script verifies both cases with an if/elif block.
Otherwise, the same error message is displayed.